// src/shared/api/types/gen/tradeRequest.ts
/**
 * Generated by orval v8.14.0 🍺
 * Do not edit manually.
 * ViaTrade API
 * API для платформы инвестиционного анализа ViaTrade
 * OpenAPI spec version: v1
 */
import type { TradeSignal } from './tradeSignal';

export type TradeRequest = {
	dateOpen: string;
	dateClose?: string;
	tradeOpen: number;
	tradeClose?: number;
	tradeSignal: TradeSignal;
	/**
	 * @minimum 0
	 * @maximum 2147483647
	 */
	count: number;
	tradeTypeId: number;
	tradeCodeId: number;
};